refactor(package): rename distribution and module to openapi-sdk
Rename the PyPI distribution from `openapi-python-sdk` to `openapi-sdk`
and the importable module from `openapi_python_sdk` to `openapi_sdk`.

- pyproject.toml: project name and packages include
- move openapi_python_sdk/ -> openapi_sdk/ (history preserved)
- update imports in examples/ and tests/
- update README.md and docs/readme-pypi.md (PyPI badges, install commands)

GitHub repository URLs and the Makefile `Project:` header are intentionally
left as `openapi-python-sdk` (the canonical repo name).

Publishing is currently blocked by PyPI's name-similarity filter; tracked
in pypi/support#9815.
