+++ docs/readme-pypi.md
@@ -1,4 +1,4 @@
-# openapi-python-sdk
+# openapi-sdk
 
 A minimal Python SDK for [Openapi®](https://openapi.it) — the largest certified API marketplace in Italy.
 Provides the core HTTP primitives to authenticate and interact with any Openapi service, without API-specific coupling.
@@ -11,7 +11,7 @@ Provides the core HTTP primitives to authenticate and interact with any Openapi 
 ## Installation
 
 ```bash
-pip install openapi-python-sdk
+pip install openapi-sdk
 ```
 
 ## Usage
@@ -21,7 +21,7 @@ Interaction with the Openapi platform happens in two distinct steps.
 ### Step 1 — Generate a token
 
 ```python
-from openapi_python_sdk.client import OauthClient
+from openapi_sdk.client import OauthClient
 
 oauth = OauthClient(username="<your_username>", apikey="<your_apikey>", test=True)
 
@@ -41,7 +41,7 @@ oauth.delete_token(id=token)
 ### Step 2 — Call an API endpoint
 
 ```python
-from openapi_python_sdk.client import Client
+from openapi_sdk.client import Client
 
 client = Client(token=token)
 
@@ -65,7 +65,7 @@ resp = client.request(
 By default, the SDK uses a 30-second timeout for all network requests. You can easily override it passing a `timeout` explicitly during initialization:
 
 ```python
-from openapi_python_sdk.client import Client
+from openapi_sdk.client import Client
 
 client = Client(token="token", timeout=60.0) # 60 seconds
 ```
