+++ tests/test_client.py
@@ -1,12 +1,12 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
-from openapi_python_sdk.client import Client, OauthClient
+from openapi_sdk.client import Client, OauthClient
 
 
 class TestOauthClient(unittest.TestCase):
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_create_token(self, mock_httpx):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"token": "abc123"}
@@ -18,7 +18,7 @@ class TestOauthClient(unittest.TestCase):
         self.assertEqual(resp["token"], "abc123")
         mock_httpx.return_value.post.assert_called_once()
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_delete_token(self, mock_httpx):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"success": True}
@@ -30,7 +30,7 @@ class TestOauthClient(unittest.TestCase):
         self.assertTrue(resp["success"])
         mock_httpx.return_value.delete.assert_called_once()
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_get_scopes(self, mock_httpx):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"scopes": ["GET:test.example.com/api"]}
@@ -41,17 +41,17 @@ class TestOauthClient(unittest.TestCase):
 
         self.assertIn("scopes", resp)
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_uses_sandbox_url_when_test_true(self, mock_httpx):
         oauth = OauthClient(username="user", apikey="key", test=True)
         self.assertIn("test.", oauth.url)
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_uses_production_url_by_default(self, mock_httpx):
         oauth = OauthClient(username="user", apikey="key")
         self.assertNotIn("test.", oauth.url)
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_auth_header_is_basic(self, mock_httpx):
         oauth = OauthClient(username="user", apikey="key")
         self.assertTrue(oauth.auth_header.startswith("Basic "))
@@ -64,7 +64,7 @@ class TestOauthClient(unittest.TestCase):
 
 class TestClient(unittest.TestCase):
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_request_get(self, mock_httpx):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"data": []}
@@ -80,7 +80,7 @@ class TestClient(unittest.TestCase):
         self.assertEqual(resp, {"data": []})
         mock_httpx.return_value.request.assert_called_once()
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_request_post(self, mock_httpx):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"result": "ok"}
@@ -95,13 +95,13 @@ class TestClient(unittest.TestCase):
 
         self.assertEqual(resp["result"], "ok")
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_auth_header(self, mock_httpx):
         client = Client(token="mytoken")
         self.assertEqual(client.auth_header, "Bearer mytoken")
         self.assertEqual(client.headers["Authorization"], "Bearer mytoken")
 
-    @patch("openapi_python_sdk.client.httpx.Client")
+    @patch("openapi_sdk.client.httpx.Client")
     def test_defaults_on_empty_request(self, mock_httpx):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {}
